feat: cleanup config and vault utils
