+++ cli/lib/crypto.ts
@@ -1,16 +1,11 @@
-import type { DropMessage } from '@shared/types/messages';
 import {
   decrypt,
-  decryptRaw,
   encrypt,
-  encryptRaw,
   hash,
-  importKeyFromBase64,
 } from '@shared/lib/crypto/operations';
-import { writeFileFromBuffer, readFileAsBuffer } from './files';
+import type { DropMessage } from '@shared/types/messages';
 import mime from 'mime';
-import { generateIV } from '@shared/lib/util';
-import { SECRET_VALUE_DELIMITER } from './constants';
+import { readFileAsBuffer, writeFileFromBuffer } from './files';
 
 export const encryptFile = async (
   key: CryptoKey,
@@ -46,26 +41,3 @@ export const hashFile = async (path: string) => {
   const fileAsArrayBuffer = await readFileAsBuffer(path);
   return hash(fileAsArrayBuffer);
 };
-
-export async function wrapSecret(key: string, value: string) {
-  const iv = generateIV();
-
-  const encryptionKey = await importKeyFromBase64(key, ['encrypt']);
-
-  const encryptedSecret = await encryptRaw(encryptionKey, iv, value);
-
-  return [iv, encryptedSecret].join(SECRET_VALUE_DELIMITER);
-}
-
-export async function unwrapSecret(
-  key: string,
-  wrappedSecret: string,
-) {
-  const [iv, encryptedValue] = wrappedSecret.split(
-    SECRET_VALUE_DELIMITER,
-  );
-
-  const decryptionKey = await importKeyFromBase64(key, ['decrypt']);
-
-  return decryptRaw(decryptionKey, iv, encryptedValue);
-}
